==> Pontia Bank S.L./1 balances.ipynb <==
import json
import pandas as pd

with open('C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Data/balances.json') as f:
    d = json.load(f)

#Transformación de los datos
data={'cliente': [], 't_id':[], 'balance_prev':[], 'balance_post':[]}

for clave in d:
    for i in range(len(d[clave])):
        data['cliente'].append(clave)
        data['t_id'].append(d[clave][i]['t_id'])
        data['balance_prev'].append(d[clave][i]['balance_prev'])
        data['balance_post'].append(d[clave][i]['balance_post'])

#print(data)

df = pd.DataFrame(data)

df.head(10)

print(df.head())

df.tail()

df.info()

#missing values
df.isna().sum()

df.isnull().sum()

#Valores únicos
df.balance_prev.value_counts()

df.balance_post.value_counts()

display(df.isna().sum(axis = 0))

df.to_csv('balance.csv', sep= '|', index=False)
