==> Pontia Bank S.L./2 tabla_clientes.ipynb <==
import json
import pandas as pd
with open("C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Data/clientes.json") as f:
    clientes=json.load(f)

lista_clientes=[]
for t_id, datos in clientes.items():
    cliente = {
        't_id': t_id,
        'origen': datos['origen'],
        'destino': datos['destino']
    }
    lista_clientes.append(cliente)

clientes=pd.DataFrame(lista_clientes)

clientes.head(10)

clientes['t_id'] = clientes['t_id'].astype(int)
clientes['origen'] = clientes['origen'].astype('string')
clientes['destino'] = clientes['destino'].astype('string')

clientes.dtypes

clientes= clientes.rename(columns={'origen': 'cl_origen','destino':'cl_destino'})

clientes.head(10)

clientes.tail()

clientes.info()

#missing values
clientes.isna().sum()

#nulos
clientes.isnull().sum()

#Valores únicos
clientes.t_id.value_counts()

clientes.cl_origen.value_counts()

clientes.cl_destino.value_counts()

clientes.to_csv('C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Csv/clientes.csv', index=False)
