==> Pontia Bank S.L./3 cuantia.ipynb <==
import json
import pandas as pd

with open('C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Data/cuantia.json') as f:
    d = json.load(f)

data={'t_id':[], 'cuantia':[]}

for clave in d:
    data['t_id'].append(clave)
    data['cuantia'].append(d[clave]['cuantia'])

df = pd.DataFrame(data)
df.head(5)

df.dtypes

df['t_id'] = df['t_id'].astype(int)

df.info()

#missing values
df.isna().sum()

df.isnull().sum()

#Valores únicos
df.t_id.value_counts()

df.to_csv('cuantia.csv', sep= '|', index=False)
