==> Pontia Bank S.L./4 tiempo.ipynb <==
import json
import pandas as pd

with open("C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Data/tiempo.json") as f:
    d=json.load(f)

data={'t_id':[], 'tiempo':[]}

for clave in d:
    for i in range(len(d[clave])):
        data['tiempo'].append(clave)
        data['t_id'].append(d[clave][i])

df = pd.DataFrame(data)
df.head()

df.tail()

df['tiempo'] = df['tiempo'].astype(int)

df.dtypes

from datetime import datetime, timedelta
start_date = datetime(2022, 9, 1, 7)
dates=[]
for i in range(1,744):
    dates.append(start_date + timedelta(hours=i-1))
date=pd.DataFrame(dates,columns=['fecha y hora'])
date.index = [i+1 for i in range(len(date))]
datefinal=date.reset_index()

datefinal.columns

nuevo_dataframe = df.merge(datefinal,left_on='tiempo', right_on='index')

dfinal=nuevo_dataframe[['t_id', 'fecha y hora']]

dfinal.dtypes

dfinal.sample(5)

dfinal.to_csv('tiempo.csv', sep= '|', index=False)
