==> Pontia Bank S.L./5 tipo.ipynb <==
import json
import pandas as pd

with open('C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Data/tipo.json') as f:
    d = json.load(f)

d.keys()

data={'t_id':[], 'tipo':[]}

for elemento in d["PAYMENT"]["t_id"]:
    data['t_id'].append(elemento)
    data['tipo'].append("PAYMENT")

for elemento in d["TRANSFER"]["t_id"]:
    data['t_id'].append(elemento)
    data['tipo'].append("TRANSFER")

for elemento in d["CASH_OUT"]["t_id"]:
    data['t_id'].append(elemento)
    data['tipo'].append("CASH_OUT")

for elemento in d["DEBIT"]["t_id"]:
    data['t_id'].append(elemento)
    data['tipo'].append("DEBIT")

for elemento in d["CASH_IN"]["t_id"]:
    data['t_id'].append(elemento)
    data['tipo'].append("CASH_IN")

df = pd.DataFrame(data)
df.head(5)

df.dtypes

df.info()

df.isna().sum()

df.isnull().sum()

df.tipo.value_counts()

df.to_csv('tipo.csv', sep= '|', index=False)
