==> Pontia Bank S.L./6 tabla_fraude.ipynb <==
import json
import pandas as pd
import numpy as np

with open("C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Data/es_fraude.json") as f:
    d1=json.load(f)

with open("C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Data/alarma_fraude.json") as f:
    d2=json.load(f)

data={'t_id':[], 'fraude':[]}

for clave in d1:
    data['t_id'].append(clave)
    data['fraude'].append(d1[clave])

df = pd.DataFrame(data)
df.head()

df.dtypes

df['t_id'] = df['t_id'].astype(int)
df['fraude'] = df['fraude'].astype('string')

df.tail()

#missing values
df.isna().sum()

df.isnull().sum()

df['fraude'].unique()

df['fraude_dummy'] = np.where( df['fraude'] == 'False' , 0, 1)

df.head()

df = df.drop('fraude', axis=1)

df.head()

df= df.rename(columns={'fraude_dummy': 'es_fraude'})

df2 = pd.DataFrame.from_dict(d2, orient="index")
#df=df.rename(columns=({'t_id':'Trans_ID','mensaje_alarma':'Alarma'}))
print(type(df2))

df2=df2.reset_index()

df2.head()

df2 = df2.drop('index', axis=1)

df2

fraude= pd.merge(df, df2 ,how='left', on='t_id')

fraude.head(10)

fraude['mensaje_alarma'] = np.where( fraude['mensaje_alarma'] == 'Detectado_fraude' , 'Detectado_fraude', 'No')

fraude.head(5)

fraude.tail()

fraude.sample(5)

fraude['mensaje_alarma'] = fraude['mensaje_alarma'].astype('string')

fraude.info()

#missing values
fraude.isna().sum()

fraude.isnull().sum()

fraude['es_fraude'].unique()

fraude['mensaje_alarma'].unique()

fraude.to_csv('C:/Users/maria/OneDrive/Escritorio/pontia/Proyecto/fraude.csv', sep='|', index=False)
