==> Pontia Bank S.L./7 tabla_transacciones.ipynb <==
#subir csv cuantía
#subir csv tiempo
#subir csv tipo

import pandas as pd
import numpy as np

cuantia = pd.read_csv('C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Csv/cuantia.csv', sep='|')

tiempo = pd.read_csv('C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Csv/tiempo.csv', sep='|')

tipo = pd.read_csv('C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Csv/tipo.csv', sep='|')

df1= cuantia.merge(tiempo ,left_on='t_id', right_on='t_id')

df1.head()

df1.tail()

df1.sample(5)

df2= df1.merge(tipo ,left_on='t_id', right_on='t_id')

df2.head()

df2.tail()

df2.sample(5)

df2= df2.rename(columns={'fecha y hora': 'tiempo'})

df2= df2.rename(columns={'cuantia': 'monto'})

df2.info()

df2['tiempo'] = df2['tiempo'].astype('datetime64')

df2['tipo'] = df2['tipo'].astype('string')

df2.info()

df2.head()

df2.to_csv('C:/Users/maria/OneDrive/Escritorio/pontia/Proyecto/transacciones.csv', sep='|', index=False)
