==> Pontia Bank S.L./8 Union_datos_maxitabla y tabla_balances.ipynb <==
#En este notebook se une toda la información en una única tabla y se crea la tabla 'balances'

import pandas as pd 
import numpy as np

clientes = pd.read_csv('C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Csv/clientes.csv')

transacciones = pd.read_csv('C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Csv/transacciones.csv')

balances = pd.read_csv('C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Csv/balance.csv')

fraude = pd.read_csv('C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Csv/fraude.csv')

clientes.head()


transacciones.sample()

fraude.sample()

balances.sample()


len(fraude)

batman=transacciones.merge(clientes, on='t_id', how='inner')

batman.head()

superman=batman.merge(fraude, on='t_id', how='inner')

superman.tail()

len(superman)

balances.head(5)

# Comprobar si hay filas en que coincidan cliente origen y destino
superman[superman['cl_origen']==superman['cl_destino']]

balances=balances.rename(columns={"cliente": "cl_origen"})
df= superman.merge(balances, how='inner', on=['cl_origen','t_id'])

df.head()

len(df)

balances=balances.rename(columns={"cl_origen": "cl_destino"})
df2= df.merge(balances, how='inner', on=['cl_destino','t_id'])

df2.head()

len(df2)

df2=df2.rename(columns={"balance_prev_x": "balance_prev_or", "balance_post_x": "balance_post_or", "balance_prev_y": "balance_prev_des", "balance_post_y": "balance_post_des"})

df2.head()

# Guardamos la tabla completa en un  nuevo csv llamado 'dec_fraude'
df2.to_csv('C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Csv/dec_fraude.csv', index=False)

balances=df2.drop(['tipo','tiempo','monto','cl_origen','cl_destino', 'es_fraude','mensaje_alarma'], axis=1, inplace=True)

df2.head()

df2.tail()

len(df2)

df2.to_csv('C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Csv/balances.csv', index=False)
