==> Pontia Bank S.L./10 Banco. Conexión BBDD a MySQL.ipynb <==
pip install pandas

from pandas.io import sql

import pandas as pd

import sqlalchemy as db

from datetime import datetime, date

import numpy as np

from scipy import stats

pip install pymysql

import pymysql

database_username='root'
database_password='vkD9aytJ'
database_ip='localhost'
database_name='banco'
database_connection=db.create_engine('mysql+pymysql://{0}:{1}@{2}/{3}'.format(database_username,database_password,database_ip,database_name))
connection=database_connection.connect()
metadata=db.MetaData()

transacciones=pd.read_csv('C:/Users/maria/OneDrive/Escritorio/pontia/Proyecto/Definitivo/transacciones.csv')

transacciones.to_sql('transaccion',connection)

clientes=pd.read_csv('C:/Users/maria/OneDrive/Escritorio/pontia/Proyecto/Definitivo/clientes.csv')

clientes.to_sql('clientes',connection)

fraude=pd.read_csv('C:/Users/maria/OneDrive/Escritorio/pontia/Proyecto/Definitivo/fraude.csv')

fraude.to_sql('fraude',connection)

balances=pd.read_csv('C:/Users/maria/OneDrive/Escritorio/pontia/Proyecto/Definitivo/balances.csv')

balances.to_sql('balances',connection)
