==> Pontia Bank S.L./14 Manipulación datos .ipynb <==
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
import numpy as np
sns.set(rc={'figure.figsize':(14,12)})

# Para suprimir los warnings de scikitlearn por convergencia prematura
from warnings import simplefilter
from sklearn.exceptions import ConvergenceWarning
simplefilter("ignore", category=ConvergenceWarning)

dec_fraude = pd.read_csv('C:/Users/Pc/Desktop/Máster/Data Science (Pontia)/Proyecto Jupyter/Csv/dec_fraude.csv')

dec_fraude.drop('Unnamed: 0', axis=1, inplace=True)

dec_fraude.head(5)

dec_fraude['tiempo'] = pd.to_datetime(dec_fraude['tiempo'])

dec_fraude['hora']=dec_fraude['tiempo'].dt.hour

dec_fraude.dtypes

# VAMOS A CREAR EL DATAFRAME PARA ENTRENAR NUESTROS MODELOS

dec_fraude.head()

dec_fraude.drop('tiempo', axis=1, inplace=True)

dec_fraude['tipo'].unique()

dec_fraude.shape

dec_fraude_tc = dec_fraude.loc[(dec_fraude['tipo'] == 'TRANSFER') | (dec_fraude['tipo'] == 'CASH_OUT')].copy()

dec_fraude_tc.reset_index(drop=True,inplace=True)

dec_fraude_tc.head()

dec_fraude_tc['tipo'].unique()

dec_fraude_tc.shape

from sklearn.preprocessing import LabelBinarizer

lb = LabelBinarizer()
dec_fraude_tc['tipo_binario'] = lb.fit_transform(dec_fraude_tc['tipo'])

dec_fraude_tc.drop('tipo', axis=1, inplace=True)

dec_fraude_tc.drop('t_id', axis=1, inplace=True)

dec_fraude_tc.drop('cl_origen', axis=1, inplace=True)

dec_fraude_tc.drop('cl_destino', axis=1, inplace=True)

dec_fraude_tc.drop('mensaje_alarma', axis=1, inplace=True)

dec_fraude_tc.drop('balance_post_or', axis=1, inplace=True)

dec_fraude_tc.drop('balance_post_des', axis=1, inplace=True)

dec_fraude_tc.head(8)

dec_fraude_tc.to_csv('C:/Users/maria/OneDrive/Escritorio/pontia/Proyecto/Definitivo/DDF.csv', sep= '|', index=False)
